requirements.txt
asgiref==3.7.2
dj-database-url==2.3.0
django==5.0.1
django-tastypie==0.15.0
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-mimeparse==1.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
django-cors-headers==4.1.0

requests==2.32.3